Add a couple of test projects for sync performance
We currently only have a project that is really small, that reflects and
average project, but we need to see the performance of larger projects
since those projects are affected in a bigger way for changes in sync
performance.

The large project consists of 21 Android and 10 java modules in a full
dependency graph with 18 variants.

Uber like project has 110 Android and 30 java modules and the
dependencies form a tree with variable width and has 4 variants.

Also reduced the number of files in average project by removing files
not required for sync.

Change-Id: Ic34787729bf67c4775a6073cd6fccc16aeca1902
Test: New tests will be generated that make use of these projects.
Bug: 113078294