Provide the ability to construct aars with a subset of resources
This CL provides the ability to construct aars that only have a smaller
set of resource files present in the res folder to actually be loaded.
This allows us to support build systems that allow specifying specific
files from a res folder to be a dependency.

As an example, if there are 100K files within a res folder, but only 1
of them need to be parsed, then the loading time of the AarSource
can be reduced from 30 seconds to 20 ms.

The change itself is fairly simple: ExternalLibrary already knows the
specific files that should be part of the library. We just change how the
ResourceSet is constructed: either we pass it the entire res folder,
or just the list of specified files.

Bug: 113121644
Test: TODO, for now just confirming that this approach is ok.
Change-Id: If5afb1e01c2690cd25ca6f340a0ff484155ad42f