Use new "ia" lib API to ensure SDK is at least 1.5
Older versions of DevMan (an APK distributed in the SDK and installed by
the CLI) are broken with APKs built using bundletool (b/111799368, b/115625864).  As part of 3.3, developers will start using bundletool by default to build instant apps. If they don't happen to be up-to-date with the
Google Play Instant SDK, things may break and they won't know why.

This change ensures that they have an SDK at least in line with the one
the API JAR expects. It also cleans up the code a bit by using the
new SdkLoader I built into the API JAR for this purpose, rather than
having to do the ServiceLoader call and version check in Studio.

I will release the new 1.5 SDK in the next day or two, well before the
canary. Here's how to point at a staged version I tested with:

  Linux:
    $ export SDK_TEST_BASE_URL=https://android-devtools-staging.corp.google.com/no_crawl/sdk_repository/tdeck-2018-09-24-19-33-72194396fc61f9565b266d653bc66b90c9698060/repository/
    $ ./bin/studio.sh
  Mac:
    $ export SDK_TEST_BASE_URL=https://android-devtools-staging.corp.google.com/no_crawl/sdk_repository/tdeck-2018-09-24-19-33-72194396fc61f9565b266d653bc66b90c9698060/repository/
    $ open /Applications/Android\ Studio.app
  Windows:
    > set SDK_TEST_BASE_URL=https://android-devtools-staging.corp.google.com/no_crawl/sdk_repository/tdeck-2018-09-24-19-33-72194396fc61f9565b266d653bc66b90c9698060/repository/
    > bin\studio[64].exe

The API and fake lib.jar were built from head @214368195

Fixes: 116539147
Test: Unit test, tested manually in upgrade scenario with staged SDK
Change-Id: Iec3feb3bad6cbd04300a3334175d3733420088bf