Rename StackedEventComponent and SimpleEventComponent
  StackedEventComponent -> ActivityComponent
  SimpleEventComponent -> EventComponent

It seems like, originally, StackedEventComponent was designed to be
a generically useful, stackable bar chart component, but in practice...

* it is only used for rendering activities and even has some activity-
  specific logic inside it
* it doesn't really support stacked bars any longer.

Therefore, we rename it based on its specific, current behavior.

Once that class was renamed, it didn't make sense anymore to distinguish
"Simple"EventComponent anymore, so we simplified / shortened its name
as well.

Bug: N/A
Test: N/A
Change-Id: I906a49bfd8d04531a0b415b195396242d5957477