Layout screen view after mode change.
A recent change on the zoom code made the method exist early if the
scale was the same, which was the case when we just change the screen
mode. And when the method exit early, it never call layoutContent() so
the screen view stays at 0,0.

This CLs add a call to layoutContent() whenever we change mode.

Test: N/A
Change-Id: I2cd5109695dc4f25c449f4f2bb9528f4b8d96c55
Fixes: 113048059