Add grey line inside the activity bar to indicate fragment changes
As we already implement the feature to show current active
Fragments, It is still hard for developers to  notice when a
fragemnt get activated/paused.
This CL add grey lines inside the activity bar to indicate
these events and the tooltip text will also change when hovering
over these lines.

Bug: 115571685
Fixes: 113512506
Test: LifecycleTooltipViewTest updated
Change-Id: I6cc3011f64ba23245635c99fd41e9b351b58ec1b