StyleModel for the intermediate shape
Here have been added classes like StyleModel that holds a BorderModel
and a FillModel. Similarly, the FillModel holds a GradientModel with an
array of GradientStopModel.

Previously, the engine would style the intermediate shape using the
SketchModel directly and modifying its properties as it ran. This CL
comes with a new separate type of object that holds information
regarding the style of the shape. The StyleModel is created using the
information in the SketchStyle object and it contains methods used to
apply transformations or transparency on the whole model without affecting
the SketchModel.

Test: DrawableFileGeneratorTest.java
Bug: N/A
Change-Id: Ia50b82b4ac2763f7e61b17e8848a27762c4f3f0f