Log blaze console output upon blaze sync failures
Currently when a UI test fails due to a bazel sync failure the test
throws a BlazeInfoException with an error code describing the failure.
This error code is not very helpful in diagnosing the root cause of such
failures.

This CL checks for blaze sync failures after every blaze sync and logs
the blaze console output to error log if it encounters a failed sync.
This should make the error logs much more useful for diagnosing bazel
project based UI tests.

Bug: 113117406
Test: Manually caused blaze sync failures and verified the console
output gets written to error logs upon the failures. Also verified that
nothing is written to error logs of the blaze sync didn't fail.

Change-Id: I71cb4dc1a756166d5195c8b82dd78e59070dbdca