Sketch converter refactoring
Removed one of the intermediate models in the converter and wrapped the
small models in a class that holds all the resources corresponding to a
page or a document (StudioResourcesModel).

Replaced AssetOptions with an interface that allows the implementation
of getters and setters for user-defined options directly in the
DrawableAsset.

This is done in order to be able to extend the code by adding different
types of assets (colors, text etc), as well as to simplify the
conversion pipeline (less levels to worry about).

There are now three levels involved in the conversion:
-> Sketch design: mimics the internal structure of the Sketch file
-> Intermediate level: converts the Sketch attributes to data that is
relevant for conversion to Studio resources
-> Studio resource: xml file(s) equivalent to the Sketch file

Bug: N/A
Test: N/A
Change-Id: I61d3c1327ee54b15d9ad109ab195cc03003b54b0