Fix problem with editing a property.
A fix for an earlier b/110880308 caused a problem with editing
values in the tabular view of the property editor.
When focus is removed we remove references to the components by
setting an EmptyProperty.
Unfortunately the fix for b/110880308 did not take this into
account. This is a fix for the regression.

Change-Id: Ie1c6922e0d992f7a009717c4e4ca7d02ab4a90d6
Fixes: b/114801903
Test: A unit test was modified to cover this case.