Replace capture chart timeline text with clock icon.
Due to the way events are managed between DurationDataRenderer and SelectionComponent (which is where custom cursors are set), MouseEvents over the label need to be consumed by DurationDataRenderer because otherwise it will be propagated to selection component which will reset the cursor.

DurationDataRenderer now does a bit of computation to determine whether the Mouse position is over a clickable label before setting the cursor.

Test: testCursorIsSetToDefaultWhenMouseIsOverLabel
Change-Id: I7855d22a2155e3b90452a8d625960f1f5e444bae
Fixes: b/75982080