Stop relying on AarSourceResourceRepository.computeResourceDirs
Stop relying on AarSourceResourceRepository.computeResourceDirs for
navigation to resource definitions. This method will be removed from
the AarSourceResourceRepository class in a subsequent CL.

Instead of creating PSI for all value XML files contained in the resource
folder of an AAR we now create PSI only for the XML files defining the
resource.

Test: existing
Bug: 74425399
Change-Id: Ia9cf9f5c07f7df066f72da5161c80fedb9d63e54