Use shortcut instead of clicking button when selecting dependencies
After https://youtrack.jetbrains.com/issue/IDEA-193540, the search
button is inlined, causing findByType(FixedSizeButton.class) to not
found the button. This problem can be the cause of other UI tests
failing since it affects all ComponentWithBrowseButton with that use
ExtendableTextComponent.

This CL uses the shortcut instead of clicking the button to fix this
issue on this particular instance.

Bug: 115387196
Test: Test now passes and manually confirm it works fine wheter the
button is inlined or not.

Change-Id: I97d8a9c188025c1af9614245f4fb8c312e2cf7de