Get layout default root tag by module
Currently, when you create a new Layout Resource Item in a module,
CreateMultiRootResourceFileAction will change the default root tag
for new layouts to ConstraintLayout if that module has constraint
layout as a dependency. The default root tag is shared by all modules,
however, and there's no code to switch it back to LinearLayout for
modules that don't depend on constraint layout, and so it's easy to get
the IDE to a state where the action recommends ConstraintLayout as the
default even though doing so would lead to compilation errors.

This CL resolves the issue by refactoring
CreateTypedResourceFileAction#getDefaultRootTag so that it accepts the
relevant module as a parameter. In CreateMultiRootResourceFileAction,
we then use the module to determine whether or not it's okay to use
ConstraintLayout as the default.

Test: CreateMultiRootResourceFileActionTest
Bug: 112834556
Change-Id: Ia89ff92aba901f491c76c665e25fd83de747dda5