[SVS] Enable Single-Variant sync by default.
With this change, all of the projects will run
single-variant sync by default, and falls back to full
variants sync only if single-variant sync is not
supported on current project.

The eligibility is detected inside of BuildAction. If
project is not eligible for single-variant sync, finish
current sync process and set a project specific flag as not
eligible for single-variant sync, start another Gradle
Sync immediately. The next Sync will be full variants
because the eligibility flag is checked at the beginning.
Eligibility flag is updated at the end of each sync.

Keep an application level flag to allow user turn off
single-variant sync globally. If this flag is set, then
Gradle sync always runs in full variants mode.

This CL also fixes several issues in new sync discovered by PSQ.
1. Remove AndroidModuleModel from facets before new sync
if it's not variant-only sync, because new sync doesn't
re-create facets anymore.
2. Check and throw injected sync errors in new sync,
for integration tests with error handlers.
3. Fix error handler to find error locations in any files.
4. Fix returned models of fetchGradleModels to include
JavaModuleModel created from jar/aar modules.
5. Fix naming of modules in new sync to handle duplicated
module names in the same project.
6. Retrieve Gradle version from wrapper file instead of
BuildScriptClasspathModel, which is not available for new sync.

Bug: 117099807
Test: included.
Change-Id: I1cea6c606234ce5b1e3141636cb57c49963ec5d6