[Build View] Add error handler for NDK incompatibility with old AGP
This changes adds a more detailed message when an error is found that commonly arises due to an incompatibility between the current version of the NDK and old plugin versions. We provide the user a quickfix to update to the latest version of the plugin which should solve this error.

Test: Unit tests added
Bug: N/A
Change-Id: I5d035ac649328410503c8e1cc76e261b620b7264