Fixing alpha comparison
Previously we inverted the value of alpha, so opaque images always
compared as identical.
We also scaled the difference between pixels by the average opacity of
the before and after. This is questionable, since while lower opacity
might result in less obvious difference for the user, it doesn't
necessarily mean that the relevant code is less incorrect.

It seems that there are a number of tests that should have been failing
but weren't because of this problem.

Test: updated existing tests
Bug: none
Change-Id: Id1217d12bcbee457e28924ce91927a5219d4c987