[CS] Fall back to normal sync for old Gradle version
When PhasedActions are not available in Gradle implementation, fall back
to normal sync.

This CL also slightly refactors the logic behind
PostSyncProjectSetup.Request.generateSourcesAfterSync: it nows keeps
this field unmodified (even for compound sync), and deals correctly with
it in SyncResultHandler.

As AGP 3.3 requires Gradle 4.10, the case where AGP supports CS but not
Gradle cannot happen.

Change-Id: Ie9f788b91393d6900cbb6ef4e9906e209762a33e
Fixes: 112021681
Test: Small logic unit tested, Gradle version manually tested as it
depends on checking in an old Gradle