Allow the package name computation to be pluggable
It is possible for the android model to place multiple resource packages
per module. For such modules, the
AndroidManifestUtils.getPackageName(facet) method can't return the correct
package name for all of them.

This patch permits the caller to supply the package computation on
construction, which permits multiple packages per module.

Bug: 110106304
Test: Existing
Change-Id: Iff9d578f1335aa830bd06e1582af9417abe1bc08