Delay the registration for project sync events.
The dependency manager for the palette is registering for project
sync events when it is created.

When a notification happens it is referencing fields that isn't
setup before the palette is shown for the first time.

This CL delays the registration such that the non initialized
fields are not being used before they are set.

Fixes: 115768776
Test: Added unit test to check project sync before setPalette()

Change-Id: I121fc1409493322923c79cb8d167c9697cfbbee7