Put recycler_view_item.xml parallel to layout file
In order to create the recycler_view_item.xml file that it needs for
switching between sample data templates, the RecyclerViewAssistant
needs to know which directory to place the file in. Currently, it
selects the first res folder given by the facet's main source provider
that can be found in the VFS. RecyclerViewAssistant's constructor
asserts that one such directory is available.

This causes problems for layouts in new resource-generating targets in
Blaze projects, which are mislabeled as belonging to the .workspace
module until a sync is performed. Since the .workspace module doesn't
have a resource directory, we see failed assertions in this scenario.

This CL side-steps the issue by having the assistant use the resource
directory that the layout file belongs to, which it finds by getting
the grandparent of the virtual file corresponding to the assistant's
component rather than going through a facet-based API.

Bug: 113346332
Test: manual

Change-Id: I5b0ce0888d900b0938c9901aaf925050e8091811