Fix race condition on Layout Editor initialization
When single variant sync is disabled, there will be two syncs when the
project is created. The first one is a "cached" one and the second, a
full sync.
The existing code in the layout editor, would wait for a sync to happen
and then for indexing to finish. Since indexing can be a lengthy
operation, this meant that, in some situations, the second full sync
would be in progress and dependencies not available.

This CL changes the start-up code for the panels so it waits until both
sync and indexing are completed.

Bug: 114307584
Test: Added test for new utility method

Change-Id: I63a9fc45c897c5ab6ff6b26e638aff8303289760