Remove Android pane after all other actions are done
IDEA now removes the Android pane from Project View if the project does
not contain any Android facet (and adds it again as soon as one is
created). But when the project is created no facets are present and this
causes to remove it if the first one created is not an Android facet.

Android Studio default pane is Android, but since the pane is removed, a
different one is selected.

This change moves the check until all pending actions are done, not when
the first facet is created.

Bug: 116737594
Test: Manual tests already cover this.
Change-Id: Ic8ca40df9e6cf05d044dc5177e8f9e06a5974709
(cherry picked from commit 9fffa01096ab334375cbec01b12b8646c95158ed)