On selection creation we force the selection componet to grab focus.
The issue is when a recording is created the selection component has focus.
The user then presses ESC to clear the selection. The selection component
consumes this event. The user presses ESC again, the selection component
has not selection and ignores this event. The event is processed by the
keyboard navigation and focus is given to the code editor.
When the user clicks on the selection component again focus is not gained
for the selection component and stays on the code editor. As such
we request focus back to the selection component when the user is interacting
with the selection component.

This change also reverts Change-Id: Iba487390d45d8f39f85028dcba3c5b7a59802916

Test: Verified manually, to verify this change three cases need to be tested.
Scenario 1.
1. Start a CPU recording
2. Stop recording by clicking the Stop button in the bottom of profiler window (not the button next to profiling config dropdown menu)
3. Press 'Esc'

Scenario 2.
1. click the clock icon in the lower-left corner of a capture. The capture will be selected.
2. Press 'Esc' twice
3. click the clock icon in the lower-left corner of a capture. The capture will be selected.
4. Press 'Esc' once

Scenario 3.
1 - Open Memory Profiler and create a heap dump
2 - Go to any other profiler (e.g. network)
3 - Use the sessions panel to navigate to the heap dump you've created (just click on it)
4 - Press ESC

Change-Id: Ic10271643f895de0dff083bbb21969de3dda30e5
Fixes: 116325518
(cherry picked from commit f385574f6f125db58c63c6afa14fae4385933d67)