Fix memory leak in tests.
IdeSdksConfigurable makes an async requests to load a list of available
SDKs/NDKs. The RepoManager does not allow canceling these requests so
they become abandoned if the configurable which is is a part of a
project service becomes unused.

This CL postpones a request to load available SDKs/NDKs until the UI is
initialized. This should slightly improve PSD startup time and fix
memory leak exceptions in tests.

Fixes: 117132146
Test: n/a
Change-Id: Id6aa3fb5502a7b88afe61637c6400aa519094ccc