Use readFromAaptNoValues in AarSourceResourceRepository
- Don't log a huge warning when constraintlayout-2.0.0 alphas
are used in a project (and they are by default in new projects)
- Carry on reading R.txt even if numeric ids don't make sense.
- Stop storing the numeric ids read from R.txt. We don't need them for
anything (the numbers are not what ends up in the APK anyway) and that's
a lot of boxed Integers potentially.
- Remove RDotTxtParser now that it's just a wrapper around SymbolIo.

Also change the repository to fall back to layout parsing if we cannot
find or parse the R.txt file at all.

Bug: 116253482
Test: AarSourceResourceRepositoryTest
Change-Id: I9e4a1fcc5ba38e831ae9be84de66a328885c4516