Show serial numbers for emulators with no avd names
When there is an emulator running without a corresponding avd
configuration, we used to show just the device properties
(manufacturer-model). This gets confusing if there are multiple
instances of emulators running with the same system image. This CL
includes the serial number in such a scenario to help distinguish
between the different emulators.

An earlier version of a similar patch only updated DeviceRenderer, but
it turns out that we have two different renderers. Future work
(b/117234701) should ideally merge the two and remove redundancy here.

Change-Id: I9c6662536bab0ad657dcc325655b12b08d30cd93
Fixes: 76116022
Test: added