Use native symbolizer for simpleperf code navigation
We were using CIDR to try to jump to code in simpleperf captures. That's not
good for three main reasons:

1) Dependency to CIDR/CLion in a repository that is pushed to AOSP;
2) The navigation logic was searching for the method name and that could
   backfire if we had a name conflict, if we don't properly parse the method
   name, etc;
3) We were navigating to the method signature's line instead of the lines
   present in the stack trace.

By using native symbolizer to do code navigation, we can safely remove the
dependency to CIDR, as it was the only place of profilers-android that
depended on it. Also, we don't rely on method name and navigate to the
correct line because native symbolizer already gives that to us.

Bug: 78459976
Bug: 117646812
Test: Added NodeNameParserTest#abiArchFileNameAndVAddressPassedToCppModel

Change-Id: I463fb79d39f2b608bcc98c3185640961557644e1