Show loading panel while Profilers is waiting for preferred process
- Updated the AndroidProfilerLaunchTaskContributor to only cache the
process run info if the profilers window is not shown. This prevents
an issue where the Profilers think it is still waiting for a preferred
process if the window is closed and re-opened.
- Commonized logic to show the loading panel for both when the profilers
is waiting for a process or for the agent to successfully attached.

Bug: 116814808
Test: Updated StudioProfilersViewTest
Change-Id: If1d0bb89a4ef246211f06ce6359e4c1bc956ab3b
(cherry picked from commit 3729a2a23a0bbb3d9907d1fe00d63f683f61f363)