[Android Plugin Upgrade] Don't upgrade composite builds.
Processing these files using the build model could yield very odd
results as we won't parse any of the parent build files and will
populate the build model with the wrong projects properties.

Also I think it should be expected that the plugin upgrade is only applied to
the current project.

Bug: N/A
Test: existing
Change-Id: I96a0122e22e20a075888b1ba388a9a53aaa59311