Resubmit "Don't create GradleFacet for wrapper root module."
Remove the creation of GradleFacet when creating top-level
module. If top-level module is a real Java or Android module,
Facets will be created during module setup.

Update comments in SyncProjectModelsSetup.

Fix ProfileRunExecutor to always display profiler icon for
Studio. The original code relies on GradleFacet to be created,
which is empty when project Window is firstly opened. This
results in profiler icon not being displayed initially for new
projects, and caused flakiness in UI tests.

Bug: 117246373
Test: TopLevelModuleFactoryTest, testNoGradleFacetInTopLevelModule.
Change-Id: I9aa23de86c6ad9ffa197ec9cd7dc57af0c081b6c