Fix NPE in BackPanel.update
Added a check whether the root of the NlComponent hierarchy
is equal to the root component of the model when determining
whether to refresh the current navigation. It seems that
the current navigation is not getting updated when
a new root is pasted into the xml file directly.

Bug:117205414

Test: None
Change-Id: I6f6b92f5f7c6e472ccc0c43b48c8015f9d2c0477
(cherry picked from commit cefb057fb83493d23192436489bd477e6aa0df44)