Add IDE support for api.jar in AARs (tools/adt/idea).
AARs may now contain a file named api.jar, used for compiling the
consumers of the AARs, whereas the traditional jars/classes.jar file is
used for packaging.

If the api.jar file is not present, we fall back to the existing
behavior: the jars/classes.jar file is used for both compiling and
packaging.

IDE support for api.jar requires changes to the model and changes to the
IDE. This CL includes changes to the model. Changes to the IDE should
follow in a separate CL.

Bug: 117145267
Test: To be included from the IDE side
Change-Id: I760107a1cac537db2d6de407b83fe2e62f0de796