Add action to re import Gradle projects
If a project is imported and some of the libraries it uses could not be
found when it is first synced, then those libraries are not searched
again unless the project is closed, removed from the recent projects and
re imported. This behaviour is in purpose, to improve performance by
looking for libraries only during first sync.

This CL adds a new action on the File menu to re import a Gradle
project. An example of how the menu looks now can be found at:

https://screenshot.googleplex.com/zUfZXjk39BG

Bug: 64143168
Test: Created tests for new action and manually confirm that the action
is present.

Change-Id: I8316759af61203770c6f2cce68d3b7321279fd55
(cherry picked from commit 9d36e3f6bf0af0a1b9a888aca30eb0c28d9df743)