Force repaint when mouseOver SelectionComponent.
Since the selection component takes the entire width of the profilers
we can't repaint on mouse enter/exit as such we need to force a repaint
on mouse move. This repaint allows the component to conditionally
draw the move bar when the mouse is over a selection area.
Without this change the SelectionComponent will not properly refresh
the entire component resulting in artifacting of the grab bar. The
area that does get refreshed is the area under the tooltip.

Test: Added test to check for repaint.
Bug: 117231848
Change-Id: I6e1ccf6fef5851b71dd665ad2e7ba89920dcc860