Update data for android-kotlin RequiresApi test
The test data was wrong, because it asserted that class literals should
be flagged for the RequiresApi Lint issue even when they are used inside an
annotation.

This change is backed by the current behavior of Lint's ApiDetector, as
well as a comment there that says, "It's okay to reference classes from
annotations." It also makes sense given that annotations are generally
processed at compile-time, thereby avoiding runtime API requirements.

Bug: 114479438
Test: //tools/adt/idea/android-kotlin/intellij.android.kotlin.tests_tests
Change-Id: Ie2777acaf3b5b05e83ba4e14cb3a639ebdcfbf86