Adds border to CPU profiling configurations combobox
We were not drawing a border around CPU profiling configurations combobox
because we are setting the JComboBox.isTableCellEditor, and IntelliJ doesn't
draw a border around comboboxes that have this property set.

JComboBox.isTableCellEditor is needed so we can be consistent with
IntelliJ's comboboxes behavior, such as not opening a dialog when navigating
over the combobox using the keyboard.

Fixes: 116814807
Test: Visual. Open CPU profiler and verify the combobox has a border around
      it. A sanity check can be made in Memory profiler as well, as its
      comboboxes should also have the border.

Change-Id: I8cf0da4e9c1c5e749e0087681a7417ca353b3c9c
(cherry picked from commit 27b5580f8120960bfb5d1526d016a7e876b8796f)