[Build View] Also display errors in the Android Gradle Plugin group.
This change ensures that lines in the output starting with error: are
displayed under the Android Gradle Plugin group. This was not added
before as I assumed that any errors would be surfaces by throwing an
exception, this was not the case. As with warnings, this will miss label
any Javac error messages that do not have a source file but this
miss labeling is better than miss labeling AGP errors as Javac ones.

Bug: N/A
Test: Unit tests added.
Change-Id: I6ebace289b5e5ee007b341f9a82cdb61b2e4f8f4