Run full-variants sync before debug NDK project.
When debugging NDK projects, the artifact is selected
based on target platform, and that artifact can be
different from the one selected in IDE.

For example, the selected variant in IDE is debug-x86,
then start debug session on an arm device, the debug
process will look for arm artifact, which might not be
synced in IDE.

Request a full-variants sync before starting the debug
session if ndk module is detected and the NdkModuleModel
has any un-synced artifacts. The target platform is
unknown to IDE before debug session is established. So
we have to sync all variants.

Add flag to GradleSyncInvoker.Request to force full-variants
sync.

Bug: 117755637
Test: current tests. Add integration test in separate cl.
Change-Id: I5780009b1a4bcd52090899185f55aeacb9789ff3
(cherry picked from commit dce99376f0e6b8d384124abc868fbbad30d3aa7e)