Adding wrong dependencies for non AndroidX project
When adding dependencies to a module, for androidx, check the
project for androidX support only. Does not matter if the
Module doesn't have app compat dependencies, if androidX is
not enabled at the project level (as there maybe other modules
or libraries in the project that have app compat dependencies).

Test: TemplateTest are still passing
Bug: 117874910
Change-Id: I1b1b33fbb6f6cbbb59981211f18e53e53497c98e