Fix FAST test: ImageAssetErrorCheckTest
Due to IJ merge, the "FixedSizeButton" cannot be clickable through
automated UI test because the "showing" value is set to false.
Fix this test issue by using different way to load local SVG file:
instead of clicking on the button, inputting the SVG file path to load
the local file.

Change-Id: I2aff4ea01161f4c3f0bcf490c56f1e40a3e6b7f7
fixes: 115539594
Test: Ran this test both with Ant and Bazel locally