Use heavy fixture in ClearResourceCacheAfterFirstBuildTest
Currently, ClearResourceCacheAfterFirstBuild has a test-only method
for resetting the project compoenent between test cases. This is
necessary because its test uses a light test fixture, meaning the
project and its components are shared between test cases. We've
recently determined that the overhead of using a heavy test fixture
is not so bad that it's worth the additional test complexity needed
to use a light fixture in such cases.

This CL updates the test to use a heavy test fixture, making the
test-only method unnecessary. Locally, I see an increased test time
of about 100ms.

Change-Id: Icaab379a28dc49a17113b726f4c583f3c113b77e
Fixes: 117988277
Test: ClearResourceCacheAfterFirstBuildTest