[New Sync] Generate unique module names.
IDEA requires module names to be unique for all of
sub-projects. This change improves deduplication
mechanism to be consistent with idea plugin.
Use ${project.name} as default, and prefixed with parts
of ${project.path} to ensure uniqeness. See
https://docs.gradle.org/current/dsl/org.gradle.plugins.ide.idea.model.IdeaModule.html#org.gradle.plugins.ide.idea.model.IdeaModule:name

This change also moves the deduplication process from
BuildAction to module setup, because module name
is a concept in the IDE side, and delimiter is defined
in project settings.

Besides, the ModuleNameGenerator supports generating
unique module names based on module path, that can
be used to guess the module names without Gradle Sync.

Bug: 117246373
Test: ModuleNameGeneratorTest.
Change-Id: I03e8fdede5702678c9d5cfc3706f0c3394f67eee