Add an hyperlink to update AGP version when build tools are not found
When sync fails due to missing build tools we offer a quickfix to
download them (given that auto download is disabled). Another way to fix
this issue is by ussing a newer AGP version that does not require buld
tools version to be specified.

This CL adds an hyperlink to update AGP version to the currently
recommended version (if possible).

Test: Modified existing tests and added test to check hyperlinks are
generated correctly based on AGP version.
Bug: 67808733
Change-Id: I5fc7569cad7209a31c175827b7c591abc5c89f27