Fix issue with removeSdkFromManifestHyperlinkTest.
This change fixes an issue whereby the contents written to the build
file was not being picked up by the ProjectBuildModel. In this change we
move from writing directly to the physical file to writing to the
Document representing the file. This ensures that the changes are
visible.

Change-Id: I2fee6f763d101fbb0d10f4eaaa6dcf7b4c08d293
Fixes: 117518708
Test: N/A
(cherry picked from commit 3f88f6678c9f5ccf3806ab40fd35f00acfb9bdaf)