Fix User Sentiment bug reporting after commit b895312
b895312 changed the code to file bug reports (via browser).
As we customized that code to take another overload to add
additional info to the bug, this code was broken in the merge.

We skipped cherry-picking that change until now. The changes
here are to make the change work along with our changes.

Change-Id: I67eea63584d2ca3b6582cd05f73ba642aa8bfe78
Fixes: 118029834
Test: manual