Change the timeline selection model.
An AccessoryPanel can now fire selection change events.

A Timeline property model can determine when a Timeline accessory
panel is active and request timeline specific objects to display
to display properties for.

The layout view property model will normally listen to selection
changes from the selection model of the surface.
When an AccessoryPanel is detected it will stop reacting to these
events, and will instead respond to selection change events from
the AccessoryPanel if there are no specific object associated with
the event.

The AccessoryPanel can therefore take over the control of the
layout view property panel.

Bug: 118388686
Test: Unit tests updated.

Change-Id: Ia4033061a6fc90808c67ea2d29d528cdaccce250