Fixing flaky testProfileOneProcessAtATime profiler test.
The issue here is we were calling response.onCompleted before
adding the process to the server state. Once onCompleted is called this frees
the main thread (our test thread) up to continue running the test.
Meanwhile the grpc thread still needs to execute the addProfiledProcess.
If this threads scheduling is delayed then the test may check the profiled
processes before this gets called resulting in an invalid result. To repo the
test failure simply revert this change and add Thread.sleep(1000) to the
addProfiledProcesss function.

Test: Fixed ;)
Change-Id: Idade5839f3851041620bad6d6e5400948c2fa220
Fixes: 110888751