Add a fast path for finding the PsiField given the field name
The default implementation in AndroidLightClassBase has two flaws:
 * It needs to construct all the PsiFields before finding
   the required one.
 * In order to identify whether a field is final or not, it relies
   on AndroidCompileUtil.findCircularDependencyOnLibraryWithSamePackage,
   which is expensive due to its use of IntelliJ's XML DOM API
   to construct Manifest classes.

As a result of these two issues, AndroidResourceUtil.findResourceFields
takes over 5 minutes sometimes in large projects (> 500 modules with
many inter dependencies). This CL adds a fast path for the specific case
of library modules where we know that the fields are all non final. In
such a case, we can get away with constructing just the requested
PsiField. As a result, the running time is ~5 seconds on the first run,
and just a few millis on subsequent invocations (compare to multiple
minutes in both scenarios earlier).

Bug: 112610798
Test: existing tests covering find usages
Change-Id: I355e22dc0133296f65a7e28e7a33ff240c14294a