Allow NewPropertyItem name lookup to have namespace spec.
Background: in the completions for a NewPropertyItem we show the
property names with a namespace prefix since the user should be
able to control exactly which property is chosen.

The TextFieldPropertyEditor will check if a value is not immediately
updated in the value before allowing the editor to move to the next
editor.

For a NewPropertyItem that causes problems when the name is edited,
since the value set: e.g. "android:text" is not the result of the
operation: name="text.

This CL allows the NewPropertyItem to test a qualified name to avoid
problems when editing a new property.

Bug: 73778668
Test: Unit test added.
Change-Id: Ibaba846f7942425189857186aea2218bf6d625b2